@dgollahon 2023-04-15
Thanks for the fast review/merge!

> Let me know if you need a release.

I'm not personally in any rush--I just happened to be following the related issue and it seemed like a fun, targeted fix.